♻️🐛 PRTL-1266 wrong link in gdcapps, make shared components
